==> 5. Attribute Recommender.ipynb <==
Jose Nico Currea, Jenna Ferguson, Jennifer Gonzalez, Evan Hadd, Muhammad Ibrahim, Ramzi Kattan, Hadley Krummel

import spacy
import pandas as pd

# Import Medium Spacy File 
nlp = spacy.load('en_core_web_md')

# Input Files
recipes = pd.read_csv('4. FeasibleRecipes.csv')
useroutput = pd.read_csv('3. UserOutput.csv')

# Attribute Column
attributes = useroutput['attributes'].dropna()

# Create Threshold for Feasibility Scores
feasible_recipes = recipes[recipes['feasibility_score'] >= 0.6]

print(f"{len(feasible_recipes)} recipes that meet your requirements found!")

# Spacy Similarity
doc1 = nlp(''.join(attributes))

# Initializing Similarity List 
spacy_scores = []

# Calculating Similarity
for text2 in feasible_recipes['review']:
    try:
        doc2 = nlp(text2)
        spacy_scores.append(doc1.similarity(doc2))
    except: 
        spacy_scores.append(0)

# Saving Review Similarity Score
feasible_recipes['Review Similarity'] = spacy_scores

# Evaluation Score
feasible_recipes['Evaluation Score'] = 0.7 *feasible_recipes['feasibility_score'] + 0.3 * feasible_recipes['Review Similarity']
feasible_recipes['Feasibility Score'] = feasible_recipes['feasibility_score']

# Finding Recommendations
recommendations = feasible_recipes[['name', 'ingredients', 'cook_time', 'rating', 'category', 'url', 'Feasibility Score', 'Review Similarity', 'Evaluation Score']].groupby(['name', 'ingredients', 'cook_time', 'rating', 'category', 'url']).mean('Evaluation Score').sort_values('Evaluation Score', ascending = False)

recommendations.head()
